backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telephony_dashboard_api"
version = "0.1.0"
description = "API for the Telephony Dashboard project."
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Edemir Junior", email = "juniormachado_@outlook.com.br" },
]
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "sqlalchemy==2.0.30",
    "psycopg2-binary==2.9.9",  # Driver para PostgreSQL
    "alembic==1.13.1",
    "pydantic==2.7.1",
    "pydantic-settings==2.2.1",
    "python-jose[cryptography]==3.3.0", # Para JWT
    "passlib[bcrypt]==1.7.4", # Para hashing de senhas
    "structlog==24.1.0",
    "httpx==0.27.0", # Para fazer chamadas à API externa
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.1",
    "pytest-cov==5.0.0",
    "pytest-mock==3.14.0",
    "pip-tools",
    "black==24.4.2", # Formatador de código
    "ruff==0.4.4",   # Linter
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = [
  "."
]